# src/models/evaluate-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * flexCharge-service
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { StringStringStringBooleanValueTuple } from './string-string-string-boolean-value-tuple';
/**
 * 
 * @export
 * @interface EvaluateResponse
 */
export interface EvaluateResponse {
    /**
     * 
     * @type {boolean}
     * @memberof EvaluateResponse
     */
    success?: boolean;
    /**
     * 
     * @type {string}
     * @memberof EvaluateResponse
     */
    result?: string | null;
    /**
     * 
     * @type {string}
     * @memberof EvaluateResponse
     */
    status?: string | null;
    /**
     * 
     * @type {string}
     * @memberof EvaluateResponse
     */
    statusCode?: string | null;
    /**
     * 
     * @type {Array<StringStringStringBooleanValueTuple>}
     * @memberof EvaluateResponse
     */
    errors?: Array<StringStringStringBooleanValueTuple> | null;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof EvaluateResponse
     */
    customProperties?: { [key: string]: string; } | null;
    /**
     * 
     * @type {string}
     * @memberof EvaluateResponse
     */
    orderSessionKey?: string | null;
    /**
     * 
     * @type {string}
     * @memberof EvaluateResponse
     */
    senseKey?: string | null;
}