# src/models/outcome-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * flexCharge-service
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { StringStringStringBooleanValueTuple } from './string-string-string-boolean-value-tuple';
/**
 * 
 * @export
 * @interface OutcomeResponse
 */
export interface OutcomeResponse {
    /**
     * 
     * @type {boolean}
     * @memberof OutcomeResponse
     */
    success?: boolean;
    /**
     * 
     * @type {string}
     * @memberof OutcomeResponse
     */
    result?: string | null;
    /**
     * 
     * @type {string}
     * @memberof OutcomeResponse
     */
    statusCode?: string | null;
    /**
     * 
     * @type {Array<StringStringStringBooleanValueTuple>}
     * @memberof OutcomeResponse
     */
    errors?: Array<StringStringStringBooleanValueTuple> | null;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof OutcomeResponse
     */
    customProperties?: { [key: string]: string; } | null;
    /**
     * 
     * @type {string}
     * @memberof OutcomeResponse
     */
    status?: string | null;
    /**
     * 
     * @type {string}
     * @memberof OutcomeResponse
     */
    confirmationId?: string | null;
}